Fix all errors that were pointed out by Lars

+++ Instance.h
@@ -16,7 +16,7 @@ using namespace std;
  *
  * @note This class defines the parameters for a Bin Packing Problem (BPP). It induces two sets:
  * - set I: containing all item indices, starting with 0 and ending in _nbItems - 1,
- * - set J: containing all bin indices, starting with 0 and ending in _nbBins
+ * - set J: containing all bin indices, starting with 0 and ending in _nbBins - 1,
  and two parameters: par_w and par_b. The read() function reads in a file with the parameter
  values, and the display() function displays the values.
  */
@@ -27,8 +27,9 @@ public:
    int _nbBins;  // Setsize of Bins (j \in J)
 
    // Parameters
-   int            par_b; // b - capacity of a single bin
+   double         par_b; // b - capacity of a single bin
    vector<double> par_w; // w_i - weight of item i \in I
+                        // dimension: _nbItems
 
    void read(string nameFile); // function to read data from a file
